Refactor user api code
